+++ releases/UEBlueprintBridge-0.5.1-UE4.24-Win64/Plugins/UEBlueprintBridge/Source/UEBlueprintBridge/UEBlueprintBridge.Build.cs
@@ -1,0 +1,12 @@
+using UnrealBuildTool;
+public class UEBlueprintBridge : ModuleRules
+{
+    public UEBlueprintBridge(ReadOnlyTargetRules Target) : base(Target)
+    {
+        PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
+        PrivateDependencyModuleNames.AddRange(new string[] {
+            "Core", "CoreUObject", "Engine", "UnrealEd", "Json", "JsonUtilities", "AssetRegistry", "BlueprintGraph", "AssetTools", "AnimGraph", "AnimGraphRuntime",
+            "MovieScene", "MovieSceneTracks", "LevelSequence", "SkeletalMeshUtilitiesCommon"
+        });
+    }
+}